Get tile maps working
Turns out instanced rendering doesn't work like I thought...
Instead, we've got a pretty weird way of doing it. We use the same quad data, and for each tile, we provide the top left texture coord. We also pass in the texture coord stride per tile as a uniform, so we can add that to the texture coordinate in the vertex shader. We have to use gl_VertexID to know which texture coordinate we should use, and we can offset this value per tile to give an easy 90 degree rotation.